qa: Use mount class in test_fscrypt to read/write and to compare trees
In TestFSCryptVolumes, make sure to use mount class to
compare directory trees and to read/write files.

Fixes: https://tracker.ceph.com/issues/74110
(cherry picked from commit aa70e125320530ceb0c80f000b6b01aed0319571)
